requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
clip==0.2.0
colorama==0.4.4
cryptography==3.4.8
easydict==1.13
fasteners==0.14.1
filelock==3.13.3
fsspec==2024.2.0
future==0.18.2
gdown==5.2.0
httplib2==0.20.2
huggingface-hub==0.25.1
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
matplotlib
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
packaging==24.0
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.3.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytz==2022.1
pyxdg==0.27
pyyaml
regex==2023.12.25
reportlab>=4.0.9
requests==2.25.1
s-tui==1.1.3
safetensors==0.4.2
six==1.16.0
soupsieve==2.6
ssh-import-id==5.11
sympy==1.12
tokenizers==0.15.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.2
transformers==4.39.1
triton==3.0.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.5
urwid==2.1.2
wadllib==1.3.6
xdg==5
zipp==1.0.0